# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "go-mqtt API",
    "version": "1.2.0",
    "description": "与当前代码实现对齐的接口文档（含 JWT 鉴权、RBAC、命令超时与重试字段）。"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/device": {
      "get": {
        "tags": ["Device"],
        "summary": "设备列表",
        "description": "RBAC: admin/operator/viewer",
        "x-roles": ["admin", "operator", "viewer"],
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeviceListEnvelope" }
              }
            }
          },
          "401": {
            "description": "未登录或 token 无效",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BaseEnvelope" }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BaseEnvelope" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Device"],
        "summary": "创建设备",
        "description": "RBAC: admin",
        "x-roles": ["admin"],
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDeviceRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeviceEnvelope" }
              }
            }
          },
          "401": {
            "description": "未登录或 token 无效",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BaseEnvelope" }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BaseEnvelope" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/device/{id}/control": {
      "post": {
        "tags": ["Device"],
        "summary": "下发控制",
        "description": "RBAC: admin/operator。响应中返回 timeout_at 和 max_retry。",
        "x-roles": ["admin", "operator"],
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [{ "$ref": "#/components/parameters/DeviceId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ControlRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ControlEnvelope" }
              }
            }
          },
          "401": {
            "description": "未登录或 token 无效",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BaseEnvelope" }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BaseEnvelope" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "DeviceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      }
    },
    "schemas": {
      "BaseEnvelope": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "data": {}
        },
        "required": ["code", "message", "data"]
      },
...
      "CreateDeviceRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "group_id": { "type": "integer" }
        }
      },
...
      "ControlRequest": {
        "type": "object",
        "required": ["cmd"],
        "properties": {
          "cmd": { "type": "string" },
          "param": {}
        }
      },
...
      "DeviceEnvelope": {
        "allOf": [{ "$ref": "#/components/schemas/BaseEnvelope" }]
      },
      "DeviceListEnvelope": {
        "allOf": [{ "$ref": "#/components/schemas/BaseEnvelope" }]
      },
      "ControlEnvelope": {
        "allOf": [{ "$ref": "#/components/schemas/BaseEnvelope" }]
      },
...
    }
  }
}